refactor!: remove support IP/RS232 fallback
This was seldom used and causes extra boilerplate for little gain.
The module will still support both RS232 and IP protocols, but will not
support automatically switching between the two.
RS232 will be the default protocol. If you want to use IP, you use the
same device name but set the ID to a socket number. You must also set
the comm mode to 'IP_DIRECT' in the online event for the virtual device
as well at setting the IP address.